add --repl option
Replace the current `command-line` spec to accept:
1. any number of input files
2. a --repl flag

If `--repl` is present, opens a `(read-eval-print-loop)` in the waffle
namespace after running each input file.

Also, this commit moves the call to `command-line` inside a `main`
submodule so that it only runs when this program is run directly (and
not when this program is required).